Move src/components/ui/theme.ts to src/theme.ts
