+++ src/components/ui/provider.tsx
@@ -1,8 +1,8 @@
 "use client";
 
+import { system } from "@/theme";
 import { ChakraProvider } from "@chakra-ui/react";
 import { ColorModeProvider, type ColorModeProviderProps } from "./color-mode";
-import { system } from "./theme";
 
 export function Provider(props: ColorModeProviderProps) {
   return (
